Update dependencies and migrate to oxc
