--- .codesandbox/ci.json
@@ -1,6 +1,0 @@
-{
-  "installCommand": "codesandbox-ci",
-  "buildCommand": "build",
-  "sandboxes": ["/example"],
-  "node": "18"
-}
