+++ .github/workflows/main.yml
@@ -6,8 +6,10 @@ jobs:
     name: Build and Typecheck
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v4
-      - uses: oven-sh/setup-bun@v1
+      - uses: actions/checkout@v6
+      - uses: oven-sh/setup-bun@v2
       - run: bun install --frozen-lockfile
       - run: bunx tsc --noEmit
-      - run: bun --bun run build
+      - run: bun run build
+      - run: bun run lint --format=github
+      - run: bun run fmt --check
