+++ .oxlintrc.json
@@ -1,0 +1,19 @@
+{
+  "$schema": "./node_modules/oxlint/configuration_schema.json",
+  "plugins": ["typescript", "unicorn", "oxc"],
+  "options": {
+    "typeAware": true,
+    "typeCheck": true,
+    "reportUnusedDisableDirectives": "warn"
+  },
+  "categories": {
+    "correctness": "error"
+  },
+  "rules": {
+    "no-explicit-any": "error",
+    "unbound-method": "allow",
+  },
+  "env": {
+    "builtin": true
+  }
+}
