+++ .vscode/settings.json
@@ -1,5 +1,5 @@
 {
   "editor.formatOnSave": true,
   "files.eol": "\n",
-  "typescript.tsdk": "node_modules/typescript/lib"
+  "js/ts.tsdk.path": "node_modules/typescript/lib"
 }
