+++ example/tsconfig.json
@@ -3,7 +3,6 @@
   "exclude": ["node_modules"],
   "compilerOptions": {
     "strict": true,
-    "esModuleInterop": false,
     "lib": ["DOM", "DOM.Iterable", "ESNext"],
     "jsx": "react-jsx",
     "target": "ESNext",
