+++ package.json
@@ -11,16 +11,16 @@
     "./package.json": "./package.json",
     ".": {
       "import": {
-        "types": "./dist/types-esm/index.d.mts",
+        "types": "./dist/react-querybuilder_chakra2.d.mts",
         "default": "./dist/react-querybuilder_chakra2.mjs"
       },
       "require": {
-        "types": "./dist/types/index.d.ts",
+        "types": "./dist/cjs/index.d.ts",
         "default": "./dist/cjs/index.js"
       }
     }
   },
-  "types": "./dist/types/index.d.ts",
+  "types": "./dist/react-querybuilder_chakra2.legacy-esm.d.ts",
   "files": [
     "dist"
   ],
@@ -46,40 +46,41 @@
   ],
   "scripts": {
     "start": "bunx --bun vite",
-    "build": "bunx --bun tsup",
+    "build": "bunx --bun tsdown",
     "typecheck": "tsc --noEmit",
     "typecheck:watch": "tsc --noEmit --watch",
-    "pretty-print": "prettier --config prettier.config.mjs --write '*.*' './src/**' './dev/**' './example/**'",
-    "pretty-check": "prettier --config prettier.config.mjs --check '*.*' './src/**' './dev/**' './example/**'",
+    "lint": "oxlint",
+    "fmt": "oxfmt",
+    "fmt:check": "oxfmt --check",
     "are-the-types-wrong": "attw --format table-flipped --pack .",
     "codesandbox-ci": "bash .codesandbox/ci.sh"
   },
   "devDependencies": {
-    "@arethetypeswrong/cli": "0.17.2",
-    "@chakra-ui/icons": "2.2.4",
-    "@chakra-ui/react": "2.10.4",
-    "@chakra-ui/system": "2.6.2",
-    "@emotion/react": "11.14.0",
-    "@emotion/styled": "11.14.0",
-    "@testing-library/dom": "10.4.0",
-    "@testing-library/jest-dom": "6.6.3",
-    "@testing-library/react": "16.1.0",
-    "@testing-library/user-event": "14.5.2",
-    "@types/bun": "1.1.16",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
-    "@types/web": "0.0.190",
-    "@vitejs/plugin-react": "4.3.4",
-    "framer-motion": "11.16.4",
-    "oxc-transform": "0.44.0",
-    "prettier": "3.4.2",
-    "react": "^19",
-    "react-dom": "^19",
-    "react-querybuilder": "8.1.0",
-    "sass": "1.83.1",
-    "tsup": "8.3.5",
-    "typescript": "5.7.3",
-    "vite": "^5"
+    "@arethetypeswrong/cli": "^0.18.2",
+    "@chakra-ui/icons": "^2.2.4",
+    "@chakra-ui/react": "^2.10.3",
+    "@chakra-ui/system": "^2.6.2",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.1",
+    "@types/bun": "^1.3.12",
+    "@types/node": "^24.0.0",
+    "@types/react": "^18.3.28",
+    "@types/react-dom": "^18.3.7",
+    "@types/web": "^0.0.345",
+    "@typescript/native-preview": "^7.0.0-dev.20260416.1",
+    "@vitejs/plugin-react": "^6.0.1",
+    "framer-motion": "^11.18.2",
+    "oxc-transform": "^0.125.0",
+    "oxfmt": "^0.45.0",
+    "oxlint": "^1.60.0",
+    "oxlint-tsgolint": "^0.21.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-querybuilder": "^7.7.1",
+    "sass": "^1.99.0",
+    "tsdown": "^0.21.9",
+    "typescript": "^6.0.2",
+    "vite": "^8.0.8"
   },
   "peerDependencies": {
     "@chakra-ui/icons": "^2",
